third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/inline-script-blocked-goofy-expected.txt

CONSOLE ERROR: The 'options' directive has been replaced with the 'unsafe-inline' and 'unsafe-eval' source expressions for the 'script-src' and 'style-src' directives. Please use those directives instead, as 'options' has no effect.
CONSOLE ERROR: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src http://127.0.0.1:*". Either the 'unsafe-inline' keyword, a hash ('sha256-791+ZkGGjayINDYVDy5oBCNVydx7XEsKAGcF1qWJq1s='), or a nonce ('nonce-...') is required to enable inline execution.

CONSOLE ERROR: Refused to execute inline event handler because it violates the following Content Security Policy directive: "script-src http://127.0.0.1:*". Either the 'unsafe-inline' keyword, a hash ('sha256-...'), or a nonce ('nonce-...') is required to enable inline execution. Note that hashes do not apply to event handlers, style attributes and javascript: navigations unless the 'unsafe-hashes' keyword is present.

This test passes if it doesn't alert fail.
